move vault mutex and refactor common CRUD funcs
