+++ internal/provider/locker.go
@@ -1,0 +1,9 @@
+package provider
+
+import "github.com/hashicorp/terraform-provider-vault/helper"
+
+// This is a global MutexKV for use within this provider.
+// Use this when you need to have multiple resources or even multiple instances
+// of the same resource write to the same path in Vault.
+// The key of the mutex should be the path in Vault.
+var VaultMutexKV = helper.NewMutexKV()
