+++ vault/resource_identity_group_member_entity_ids.go
@@ -1,22 +1,18 @@
 package vault
 
 import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
 	"github.com/hashicorp/terraform-provider-vault/internal/consts"
 	"github.com/hashicorp/terraform-provider-vault/internal/identity/group"
-	"github.com/hashicorp/terraform-provider-vault/internal/provider"
 )
 
 func identityGroupMemberEntityIdsResource() *schema.Resource {
 	return &schema.Resource{
-		CreateContext: identityGroupMemberEntityIdsUpdate,
-		UpdateContext: identityGroupMemberEntityIdsUpdate,
-		ReadContext:   ReadContextWrapper(identityGroupMemberEntityIdsRead),
-		DeleteContext: identityGroupMemberEntityIdsDelete,
+		CreateContext: group.GetGroupMemberUpdateContextFunc(group.EntityResourceType),
+		UpdateContext: group.GetGroupMemberUpdateContextFunc(group.EntityResourceType),
+		ReadContext:   ReadContextWrapper(group.GetGroupMemberReadContextFunc(group.EntityResourceType, true)),
+		DeleteContext: group.GetGroupMemberDeleteContextFunc(group.EntityResourceType),
 
 		Schema: map[string]*schema.Schema{
 			consts.FieldMemberEntityIDs: {
@@ -49,45 +45,4 @@ use "data.vault_identity_group.*.group_name", "vault_identity_group.*.group_name
 			},
 		},
 	}
-}
-
-func identityGroupMemberEntityIdsUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	gid := d.Get(consts.FieldGroupID).(string)
-	path := group.IdentityGroupIDPath(gid)
-	vaultMutexKV.Lock(path)
-	defer vaultMutexKV.Unlock(path)
-
-	client, e := provider.GetClient(d, meta)
-	if e != nil {
-		return diag.FromErr(e)
-	}
-
-	if diag := group.UpdateGroupMemberContextFunc(d, client, consts.FieldMemberEntityIDs); diag != nil {
-		return diag
-	}
-
-	return identityGroupMemberEntityIdsRead(ctx, d, meta)
-}
-
-func identityGroupMemberEntityIdsRead(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	client, e := provider.GetClient(d, meta)
-	if e != nil {
-		return diag.FromErr(e)
-	}
-
-	return group.ReadGroupMemberContextFunc(d, client, consts.FieldMemberEntityIDs, true)
-}
-
-func identityGroupMemberEntityIdsDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	id := d.Get(consts.FieldGroupID).(string)
-	path := group.IdentityGroupIDPath(id)
-	vaultMutexKV.Lock(path)
-	defer vaultMutexKV.Unlock(path)
-
-	client, e := provider.GetClient(d, meta)
-	if e != nil {
-		return diag.FromErr(e)
-	}
-
-	return group.DeleteGroupMemberContextFunc(d, client, consts.FieldMemberEntityIDs)
 }
