+++ website/vault.erb
@@ -289,6 +289,10 @@
                             <a href="/docs/providers/vault/r/identity_group_member_entity_ids.html">vault_identity_group_member_entity_ids</a>
                         </li>
 
+                        <li<%= sidebar_current("docs-vault-resource-identity-group-member-group-ids") %>>
+                            <a href="/docs/providers/vault/r/identity_group_member_group_ids.html">vault_identity_group_member_group_ids</a>
+                        </li>
+
                         <li<%= sidebar_current("docs-vault-resource-identity-group-policies") %>>
                             <a href="/docs/providers/vault/r/identity_group_policies.html">vault_identity_group_policies</a>
                         </li>
